+++ SBBMobileDesignSystemSwiftUI/Views/SBBCheckBox.swift
@@ -7,10 +7,7 @@ import SwiftUI
 public extension SBBCheckBox where Content == EmptyView {
     init(isOn: Binding<Bool>, image: Image? = nil, label: String, showTextFieldLine: Bool = true) {
         self._isOn = isOn
-        self.image = image
-        self.label = LocalizedStringKey(label)
-        self.showTextFieldLine = showTextFieldLine
-        self.content = nil
+        self.checkboxAndRadioButtonContainer = CheckBoxAndRadioButtonContainer(type: .checkbox, isOn: isOn, image: image, label: label, showTextFieldLine: showTextFieldLine)
     }
 }
 
@@ -18,72 +15,16 @@ public struct SBBCheckBox<Content>: View where Content: View {
     
     @Environment(\.isEnabled) private var isEnabled
     @Binding private var isOn: Bool
-    private var image: Image?
-    private var label: LocalizedStringKey?
-    private let content: Content?
-    private let showTextFieldLine: Bool
     
-    private var icon: Image {
-        switch (isOn, isEnabled) {
-        case (true, true):
-            return Image("Checked", bundle: Helper.bundle)
-        case (true, false):
-            return Image("Checked_disabled", bundle: Helper.bundle)
-        case (false, true):
-            return Image("Unchecked", bundle: Helper.bundle)
-        case (false, false):
-            return Image("Unchecked_disabled", bundle: Helper.bundle)
-        }
-    }
+    private var checkboxAndRadioButtonContainer: CheckBoxAndRadioButtonContainer<Content>
     
     public init(isOn: Binding<Bool>, showTextFieldLine: Bool = false, @ViewBuilder content: @escaping () -> Content) {
         self._isOn = isOn
-        self.image = nil
-        self.label = nil
-        self.showTextFieldLine = showTextFieldLine
-        self.content = content()
+        self.checkboxAndRadioButtonContainer = CheckBoxAndRadioButtonContainer(type: .checkbox, isOn: isOn, showTextFieldLine: showTextFieldLine, content: content)
     }
     
     public var body: some View {
-        ZStack(alignment: .bottom) {
-            HStack(alignment: .top, spacing: 8) {
-                icon
-                    .accessibility(hidden: true)
-                if image != nil {
-                    image!
-                        .resizeToContentSizeCategory(originalHeight: 24)
-                        .accessibility(hidden: true)
-                }
-                if label != nil {
-                    Text(label!)
-                        .sbbFont(.body)
-                        .padding(.top, 2)
-                        .fixedSize(horizontal: false, vertical: true)
-                }
-                if content != nil {
-                    content!
-                }
-                Spacer()
-            }
-                .padding(.vertical, 12)
-                .padding(.horizontal, 16)
-            if showTextFieldLine {
-                Rectangle()
-                    .fill(Color.sbbColorInternal(.textfieldLineInactive))
-                    .frame(height: 1)
-                    .padding(.leading, 16)
-            }
-        }
-            .background(Color.clear)
-            .onTapGesture {
-                self.isOn.toggle()
-            }
-            .accessibilityElement(children: .combine)
-            .accessibility(addTraits: .isButton)
-            .accessibility(value: isOn ? Text("\("ON".localized)") : Text("\("OFF".localized)"))
-            .accessibilityAction {
-                self.isOn.toggle()
-            }
+        checkboxAndRadioButtonContainer
     }
 }